requirements.txt
Flask==3.0.0
Authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
